+++ app/controllers/lens_controller.rb
@@ -17,6 +17,7 @@ class LensController < ApplicationController
   end
 
   post "/lens" do
+    redirect_if_not_logged_in
     len = Len.new(
       make: params[:len][:make],
       model: params[:len][:model],
@@ -64,31 +65,39 @@ class LensController < ApplicationController
   end
 
   patch "/lens/:id" do
-    len = Len.find_by(id: params[:id])
-    len.update(
-      make: params[:len][:make],
-      model: params[:len][:model],
-      focal_range: params[:len][:focal_range],
-      min_aperture: params[:len][:min_aperture],
-      max_aperture: params[:len][:max_aperture],
-      camera_id: params[:len][:camera_id],
-      user_id: params[:len][:user_id],
-      updated_at: Time.now
-    )
-    if len.save
-      redirect to "/lens/#{len.id}"
+    redirect_if_not_logged_in
+    if len = Len.find_by(id: params[:id])
+      len.update(
+        make: params[:len][:make],
+        model: params[:len][:model],
+        focal_range: params[:len][:focal_range],
+        min_aperture: params[:len][:min_aperture],
+        max_aperture: params[:len][:max_aperture],
+        camera_id: params[:len][:camera_id],
+        user_id: params[:len][:user_id],
+        updated_at: Time.now
+      )
+      if len.save
+        redirect to "/lens/#{len.id}"
+      else
+        @errors = len.errors.full_messages
+        @len = len
+        erb :"/lens/edit"
+      end
     else
-      @errors = len.errors.full_messages
-      @len = len
-      erb :"/lens/edit"
+      not_authorized
     end
   end
 
   delete "/lens/:id" do
-    len = Len.find_by(id: params[:id])
-    flash[:message] = "#{len.make} - #{len.model} deleted successfully."
-    len.destroy
-    redirect to "/lens"
+    redirect_if_not_logged_in
+    if len = Len.find_by(id: params[:id])
+      flash[:message] = "#{len.make} - #{len.model} deleted successfully."
+      len.destroy
+      redirect to "/lens"
+    else
+      not_authorized
+    end
   end
   
 end
